llm_mcp/pdf-maker/requirements.txt
mcp>=1.0.0
json-repair>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
pytest>=8.0.0
uvicorn[standard]>=0.29.0
anyio>=4.0.0
